requirements.txt
Flask==2.3.2
Pillow==9.4.0
numpy==1.25.2
